Fixed bug where the value at a position in Big_Number was multiplied by the multiplicand as an int (instead of long long int) which could cause an integer overflow. Updated with test in main.c to check it. Need to fix more edge cases and make tests for formalized.

+++ Big_Numbers/Big_Numbers.c
@@ -99,7 +99,7 @@ void mulIntToBig(Big_Number* bigNumber, int number){
 
 		//Compute the new value for this position
 		//and hold it in temp to prevent overflow
-		temp = (bigNumber->number[cur] * number) + remainder;
+		temp = ((long long int)(bigNumber->number[cur]) * number) + remainder;
 
 		//If temp > what can be held in a position
 		if (temp > maxbitValue){
